+++ DesignToolbox/DesignToolbox/Utils/A11YIdentifiers.swift
@@ -24,4 +24,9 @@ enum A11YIdentifiers {
     static let componentRadioItem = "designSystemToolbox_component_radioItem"
     static let componentSwitch = "designSystemToolbox_component_switch"
     static let componentSwitchItem = "designSystemToolbox_component_switchitem"
+
+    // MARK: - Configurations
+
+    static let configurationTextField = "designSystemToolbox_configuration_textField"
+    static let configurationSwitchSelection = "designSystemToolbox_configuration_switch_selection"
 }